# adaptive batching