# What does this PR do?

This PR implement the adaptive batching which means that the max total tokens and batch size will change according the requests.

Our customer need the configuration that the MXA_TOTAL_TOKENS is 8K and MAX_INPUT_TOKENS is 4K. This means that the customer need the 8K max total token capability. But most of the request's total tokens is within 2K or 4K. If we use the actual 8K max total token. Then we can just set the batch size to 32. If we set a bigger batch size then OOM will occur. This lead to a bad performance for many benchmarks, for example the vllm benchmark. If we use adaptive batching then we can use a smaller actual max total token and set a bigger batch size. This will give a better performance. 

We use the following benchmark to test.

model=meta-llama/Meta-Llama-3-8B-Instruct
~/vllm/benchmarks# python3 benchmark_serving.py --base-url=http://127.0.0.1:8088/ --backend vllm  --dataset-name sharegpt --request-rate 10 --model $model --dataset-path

If we don't use the adaptive batching we get the following benchmark result.

_Traffic request rate: 10.0
100%
============ Serving Benchmark Result ============
Successful requests:                     256
Benchmark duration (s):                  193.11
Total input tokens:                      56983
Total generated tokens:                  165271
Request throughput (req/s):              1.33
Input token throughput (tok/s):          295.08
**Output token throughput (tok/s):         855.82**
---------------Time to First Token----------------
Mean TTFT (ms):                          61144.68
Median TTFT (ms):                        59155.51
P99 TTFT (ms):                           135269.15
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          31.77
Median TPOT (ms):                        32.87
P99 TPOT (ms):                           43.40
==================================================_

If we use the adaptive batching we get the following benchmark result.

_Traffic request rate: 10.0
100%
============ Serving Benchmark Result ============
Successful requests:                     256
Benchmark duration (s):                  101.17
Total input tokens:                      56983
Total generated tokens:                  163409
Request throughput (req/s):              2.53
Input token throughput (tok/s):          563.22
**Output token throughput (tok/s):         1615.14**
---------------Time to First Token----------------
Mean TTFT (ms):                          25545.90
Median TTFT (ms):                        23467.30
P99 TTFT (ms):                           67305.21
-----Time per Output Token (excl. 1st token)------
Mean TPOT (ms):                          23.46
Median TPOT (ms):                        23.36
P99 TPOT (ms):                           54.14
==================================================_

In this PR we introduced a parameter MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST and this parameter define the {max total tokens, max batch size} key value pair.

For example, if you use MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST=2048:48,4096:24,8192:12.
Then if all the request with total token <= 2048 the max batch size will be 48. Also actual running batch size changes according to the max total tokens in the batch. 

If you remove the environment variable MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST the old-style batching will be used.

Did you write any new necessary tests?
I have test this PR in vllm benchmark with many models and data type.

## Who can review?

Anyone in the community is free to review the PR once the tests have passed. Feel free to tag
members/contributors who may be interested in your PR.

<!-- Your PR will be replied to more quickly if you can figure out the right person to tag with @


@OlivierDehaene OR @Narsil

 -->