@kdamaszk 2024-07-23
@donzh please fix rust tests (`cargo test`).