@donzh 2024-07-25
> @donzh Could it possible that you add some unit test to test this adaptive batching functionality? For example in test_causal_lm.py ?

@jczaja Actually I checked the unit test in test_causal_lm.py when I developing adaptive batching. The unit test for adaptive batching feature that can be done in test_causal_lm.py is very limited. The adaptive batching feature need the rust code(infer.rs) and Python code (causal_lm.py) work together. If I want to just test lm part of the adaptive batching feature I need to set the BATCH_BUCKET_SIZE to very small (maybe 2) and many other parameters.  Is there any method to change the these environment parameters in test_causal_lm.py? For example, Change the BATCH_BUCKET_SIZE to 2 MAX_TOTAL_TOKENS to 128 and other parameters. If this can be done a unit test for function "recombine" can be added. But this is also a very trial test.