@donzh 2024-07-25
I drew a picture to illustrate the adaptive batching.

![image](https://github.com/user-attachments/assets/e4dcf960-b2a8-4a4e-8915-0b6924e5add0)