@donzh 2024-08-06
> @donzh , can you add explanation of adaptive batching to README and add a command example? thanks

update README and the parameter MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST is a "add -e in docker run command" parameter. The following config is used in my test.

#!/bin/bash

#--model-id
export MODEL_ID=meta-llama/Meta-Llama-3-8B-Instruct
#--port
export PORT=8089

export RUST_BACKTRACE=full

#enable trace
export DBG_TRACE_FILENAME=./dbg_trace.log
rm dbg_trace.log

# Gaudi setup
export LIMIT_HPU_GRAPH=True
export ENABLE_HPU_GRAPH=True

export MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST=2048:48,4096:24,8192:12

# TGI server configs
#--max-input-tokens
export MAX_INPUT_TOKENS=4096
#--max-total-tokens
export MAX_TOTAL_TOKENS=8192
#export WARMUP_ENABLED=false
# generate setup
export BATCH_BUCKET_SIZE=12
#--max-batch-size
export MAX_BATCH_SIZE=12
#--max-batch-total-tokens
export MAX_BATCH_TOTAL_TOKENS=$(( MAX_BATCH_SIZE*MAX_TOTAL_TOKENS))

#prefill setup
#--max-batch-prefill-tokens
export MAX_BATCH_PREFILL_TOKENS=$(( MAX_INPUT_TOKENS * 12 ))
export PREFILL_BATCH_BUCKET_SIZE=4
export PAD_SEQUENCE_TO_MULTIPLE_OF=128

#general server configs
#--max-concurrent-requests
export MAX_CONCURRENT_REQUESTS=512
#--waiting-served-ratio
export WAITING_SERVED_RATIO=0.3
#--sharded
export SHARDED=false