@donzh 2024-08-06
> This PR is on the code changes for TGI. why we show the benefits to vllm?

The benefit is to TGI not to vllm. We just use vllm benchmark to test.
The reason it that our customer care about vllm benchmark and compare Gaudi's result with A100's result. I will test the TGI benchmark.