@donzh 2024-08-06
The TGI benchmark test results for 8K total and 4K input are as follows. The throughput increases about 60% and TTFT is decrease about 30%-40% with adaptive batch.

Without adaptive batch (MAX_BATCH_SIZE=32):

| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 42.07 ms  | 41.35 ms  | 43.72 ms  | 41.82 ms  | 43.72 ms  | 43.72 ms  |
|                | 2          | 41.81 ms  | 40.86 ms  | 43.20 ms  | 41.87 ms  | 43.20 ms  | 43.20 ms  |
|                | 4          | 40.89 ms  | 40.29 ms  | 41.82 ms  | 40.65 ms  | 41.82 ms  | 41.82 ms  |
|                | 8          | 41.92 ms  | 41.53 ms  | 43.15 ms  | 41.78 ms  | 43.15 ms  | 43.15 ms  |
|                | 16         | 41.82 ms  | 40.64 ms  | 42.62 ms  | 41.95 ms  | 42.62 ms  | 42.62 ms  |
|                | 32         | 42.24 ms  | 41.01 ms  | 43.34 ms  | 42.24 ms  | 43.34 ms  | 43.34 ms  |
| Decode (token) | 1          | 28.03 ms  | 28.00 ms  | 28.07 ms  | 28.03 ms  | 28.03 ms  | 28.03 ms  |
|                | 2          |           |           | 28.05 ms  |           | 28.02 ms  | 28.02 ms  |
|                | 4          | 28.05 ms  | 28.03 ms  | 28.11 ms  | 28.05 ms  | 28.11 ms  | 28.11 ms  |
|                | 8          | 29.02 ms  | 29.00 ms  | 29.05 ms  | 29.02 ms  | 29.02 ms  | 29.02 ms  |
|                | 16         | 30.31 ms  | 30.30 ms  | 30.33 ms  | 30.31 ms  | 30.31 ms  | 30.31 ms  |
|                | 32         | 29.68 ms  | 29.66 ms  | 29.71 ms  | 29.68 ms  | 29.71 ms  | 29.71 ms  |
| Decode (total) | 1          | 224.24 ms | 224.03 ms | 224.54 ms | 224.25 ms | 224.23 ms | 224.23 ms |
|                | 2          | 224.22 ms | 224.04 ms | 224.41 ms | 224.22 ms | 224.19 ms | 224.19 ms |
|                | 4          | 224.44 ms | 224.22 ms | 224.85 ms | 224.42 ms | 224.85 ms | 224.85 ms |
|                | 8          | 232.13 ms | 231.96 ms | 232.42 ms | 232.13 ms | 232.13 ms | 232.13 ms |
|                | 16         | 242.45 ms | 242.37 ms | 242.64 ms | 242.46 ms | 242.51 ms | 242.51 ms |
|                | 32         | 237.41 ms | 237.27 ms | 237.66 ms | 237.41 ms | 237.66 ms | 237.66 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 23.78 tokens/secs   | 22.87 tokens/secs   | 24.19 tokens/secs   |
|         | 2          | 47.85 tokens/secs   | 46.30 tokens/secs   | 48.95 tokens/secs   |
|         | 4          | 97.85 tokens/secs   | 95.65 tokens/secs   | 99.29 tokens/secs   |
|         | 8          | 190.87 tokens/secs  | 185.40 tokens/secs  | 192.63 tokens/secs  |
|         | 16         | 382.70 tokens/secs  | 375.44 tokens/secs  | 393.73 tokens/secs  |
|         | 32         | 757.78 tokens/secs  | 738.41 tokens/secs  | 780.26 tokens/secs  |
| Decode  | 1          | 35.68 tokens/secs   | 35.63 tokens/secs   | 35.71 tokens/secs   |
|         | 2          | 71.36 tokens/secs   | 71.30 tokens/secs   | 71.41 tokens/secs   |
|         | 4          | 142.58 tokens/secs  | 142.32 tokens/secs  | 142.72 tokens/secs  |
|         | 8          | 275.71 tokens/secs  | 275.37 tokens/secs  | 275.91 tokens/secs  |
|         | 16         | 527.93 tokens/secs  | 527.52 tokens/secs  | 528.11 tokens/secs  |
|         | 32         | 1078.29 tokens/secs | 1077.16 tokens/secs | 1078.95 tokens/secs |

With adaptive batch:


| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 41.27 ms  | 40.47 ms  | 42.25 ms  | 41.04 ms  | 42.25 ms  | 42.25 ms  |
|                | 2          | 41.18 ms  | 40.54 ms  | 41.97 ms  | 41.12 ms  | 41.97 ms  | 41.97 ms  |
|                | 4          | 41.11 ms  | 40.61 ms  | 42.06 ms  | 40.93 ms  | 42.06 ms  | 42.06 ms  |
|                | 8          | 41.55 ms  | 40.70 ms  | 42.38 ms  | 41.91 ms  | 42.38 ms  | 42.38 ms  |
|                | 16         | 41.83 ms  | 40.89 ms  | 42.85 ms  | 42.03 ms  | 42.85 ms  | 42.85 ms  |
|                | 32         | 41.90 ms  | 41.10 ms  |           | 41.84 ms  |           |           |
| Decode (token) | 1          | 18.97 ms  | 18.85 ms  | 19.17 ms  | 18.98 ms  | 18.91 ms  | 18.91 ms  |
|                | 2          | 18.95 ms  | 18.84 ms  | 19.11 ms  | 18.93 ms  | 18.84 ms  | 18.84 ms  |
|                | 4          | 19.01 ms  | 18.89 ms  | 19.09 ms  | 19.05 ms  | 19.03 ms  | 19.03 ms  |
|                | 8          |           | 18.90 ms  | 19.10 ms  | 19.06 ms  |           |           |
|                | 16         | 19.11 ms  | 18.98 ms  | 19.30 ms  | 19.12 ms  | 19.08 ms  | 19.08 ms  |
|                | 32         | 19.09 ms  | 19.00 ms  | 19.22 ms  | 19.11 ms  | 19.22 ms  | 19.22 ms  |
| Decode (total) | 1          | 151.78 ms | 150.78 ms | 153.39 ms | 151.83 ms | 151.24 ms | 151.24 ms |
|                | 2          | 151.61 ms | 150.75 ms | 152.85 ms | 151.43 ms | 150.75 ms | 150.75 ms |
|                | 4          | 152.06 ms | 151.08 ms | 152.75 ms | 152.39 ms | 152.23 ms | 152.23 ms |
|                | 8          | 152.05 ms | 151.19 ms | 152.77 ms | 152.50 ms | 152.24 ms | 152.24 ms |
|                | 16         | 152.88 ms | 151.80 ms | 154.37 ms | 152.94 ms | 152.62 ms | 152.62 ms |
|                | 32         | 152.75 ms | 152.00 ms | 153.75 ms | 152.88 ms | 153.75 ms | 153.75 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 24.24 tokens/secs   | 23.67 tokens/secs   | 24.71 tokens/secs   |
|         | 2          | 48.57 tokens/secs   | 47.66 tokens/secs   | 49.33 tokens/secs   |
|         | 4          | 97.30 tokens/secs   | 95.11 tokens/secs   | 98.50 tokens/secs   |
|         | 8          | 192.57 tokens/secs  | 188.76 tokens/secs  | 196.57 tokens/secs  |
|         | 16         | 382.62 tokens/secs  | 373.36 tokens/secs  | 391.30 tokens/secs  |
|         | 32         | 763.93 tokens/secs  | 746.77 tokens/secs  | 778.59 tokens/secs  |
| Decode  | 1          | 52.71 tokens/secs   | 52.15 tokens/secs   | 53.06 tokens/secs   |
|         | 2          | 105.54 tokens/secs  | 104.68 tokens/secs  | 106.13 tokens/secs  |
|         | 4          | 210.45 tokens/secs  | 209.49 tokens/secs  | 211.80 tokens/secs  |
|         | 8          | 420.92 tokens/secs  | 418.94 tokens/secs  | 423.32 tokens/secs  |
|         | 16         | 837.28 tokens/secs  | 829.17 tokens/secs  | 843.19 tokens/secs  |
|         | 32         | 1676.01 tokens/secs | 1665.06 tokens/secs | 1684.17 tokens/secs |