@tthakkal 2024-08-08
@donzh 

I tested this PR with 1.17 synapseAI 
used docker image `vault.habana.ai/gaudi-docker/1.17.0/ubuntu22.04/habanalabs/pytorch-installer-2.3.1:latest` as base, latest changes from OH and deepspeed 1.17, attached patch file with same changes here
[docker_changes.patch](https://github.com/user-attachments/files/16551927/docker_changes.patch)

Decode performance is less when we use adaptive batching . not much improvement on prefill. could you please verify.


## Without adaptive batching 
docker command 
```
docker run --rm -p 8080:80 \
   --runtime=habana \
   -v <data>:/data \
   -e HUGGING_FACE_HUB_TOKEN=<token> \
   -e HABANA_VISIBLE_DEVICES=7 \
   -e OMPI_MCA_btl_vader_single_copy_mechanism=none \
   -e HF_HUB_ENABLE_HF_TRANSFER=1 \
   -e PREFILL_BATCH_BUCKET_SIZE=1 \
   -e BATCH_BUCKET_SIZE=32 \
   -e PAD_SEQUENCE_TO_MULTIPLE_OF=128 \
   -e ENABLE_HPU_GRAPH=true \
   -e LIMIT_HPU_GRAPH=true \
   -e USE_FLASH_ATTENTION=true \
   -e FLASH_ATTENTION_RECOMPUTE=true \
   --cap-add=sys_nice \
   --ipc=host \
   --name tgi_gaudi_adaptive \
   tgi_gaudi_adaptive \
   --model-id meta-llama/Meta-Llama-3-8B-Instruct \
   --max-input-tokens 1024 \
   --max-total-tokens 2048 \
   --max-batch-size 32
```

benchmark results:

| Parameter          | Value                               |
|--------------------|-------------------------------------|
| Model              | meta-llama/Meta-Llama-3-8B-Instruct |
| Sequence Length    | 10                                  |
| Decode Length      | 8                                   |
| Top N Tokens       | None                                |
| N Runs             | 10                                  |
| Warmups            | 1                                   |
| Temperature        | None                                |
| Top K              | None                                |
| Top P              | None                                |
| Typical P          | None                                |
| Repetition Penalty | None                                |
| Frequency Penalty  | None                                |
| Watermark          | false                               |
| Do Sample          | false                               |


| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 42.36 ms  | 40.61 ms  | 46.25 ms  | 41.98 ms  | 46.25 ms  | 46.25 ms  |
|                | 2          | 43.56 ms  | 40.24 ms  | 48.52 ms  | 42.55 ms  | 48.52 ms  | 48.52 ms  |
|                | 4          | 42.90 ms  | 41.67 ms  | 50.17 ms  | 42.27 ms  | 50.17 ms  | 50.17 ms  |
|                | 8          | 41.75 ms  | 40.09 ms  | 43.67 ms  | 41.20 ms  | 43.67 ms  | 43.67 ms  |
|                | 16         | 43.66 ms  | 40.20 ms  | 61.08 ms  | 41.61 ms  | 61.08 ms  | 61.08 ms  |
|                | 32         | 45.22 ms  | 41.54 ms  | 55.24 ms  | 42.73 ms  | 55.24 ms  | 55.24 ms  |
| Decode (token) | 1          | 8.77 ms   | 8.69 ms   | 8.91 ms   | 8.81 ms   | 8.70 ms   | 8.70 ms   |
|                | 2          | 8.81 ms   | 8.70 ms   | 8.90 ms   | 8.82 ms   | 8.79 ms   | 8.79 ms   |
|                | 4          | 8.75 ms   |           | 8.82 ms   | 8.76 ms   | 8.70 ms   | 8.70 ms   |
|                | 8          | 8.88 ms   | 8.83 ms   | 8.93 ms   | 8.89 ms   | 8.86 ms   | 8.86 ms   |
|                | 16         | 10.73 ms  | 10.67 ms  | 10.85 ms  | 10.72 ms  | 10.76 ms  | 10.76 ms  |
|                | 32         | 13.22 ms  | 13.14 ms  | 13.46 ms  | 13.19 ms  | 13.46 ms  | 13.46 ms  |
| Decode (total) | 1          | 70.15 ms  | 69.55 ms  | 71.25 ms  | 70.51 ms  | 69.59 ms  | 69.59 ms  |
|                | 2          | 70.48 ms  | 69.57 ms  | 71.21 ms  | 70.60 ms  | 70.35 ms  | 70.35 ms  |
|                | 4          | 70.00 ms  | 69.56 ms  | 70.56 ms  | 70.05 ms  | 69.56 ms  | 69.56 ms  |
|                | 8          | 71.04 ms  | 70.66 ms  | 71.42 ms  | 71.13 ms  | 70.92 ms  | 70.92 ms  |
|                | 16         | 85.82 ms  | 85.35 ms  | 86.78 ms  | 85.74 ms  | 86.10 ms  | 86.10 ms  |
|                | 32         | 105.80 ms | 105.14 ms | 107.69 ms | 105.48 ms | 107.69 ms | 107.69 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 23.64 tokens/secs   | 21.62 tokens/secs   | 24.62 tokens/secs   |
|         | 2          | 46.07 tokens/secs   | 41.22 tokens/secs   | 49.70 tokens/secs   |
|         | 4          | 93.49 tokens/secs   | 79.73 tokens/secs   | 96.00 tokens/secs   |
|         | 8          | 191.79 tokens/secs  | 183.19 tokens/secs  | 199.54 tokens/secs  |
|         | 16         | 371.49 tokens/secs  | 261.95 tokens/secs  | 397.99 tokens/secs  |
|         | 32         | 715.49 tokens/secs  | 579.32 tokens/secs  | 770.27 tokens/secs  |
| Decode  | 1          | 114.05 tokens/secs  | 112.28 tokens/secs  | 115.03 tokens/secs  |
|         | 2          | 227.01 tokens/secs  | 224.69 tokens/secs  | 229.98 tokens/secs  |
|         | 4          | 457.12 tokens/secs  | 453.52 tokens/secs  | 460.03 tokens/secs  |
|         | 8          | 900.91 tokens/secs  | 896.16 tokens/secs  | 905.74 tokens/secs  |
|         | 16         | 1491.58 tokens/secs | 1474.89 tokens/secs | 1499.65 tokens/secs |
|         | 32         | 2419.87 tokens/secs | 2377.14 tokens/secs | 2434.91 tokens/secs |
'



## With Adaptive batching
docker command

```
docker run --rm -p 8080:80 \
   --runtime=habana \
   -v <data>:/data \
   -e HUGGING_FACE_HUB_TOKEN=<token> \
   -e HABANA_VISIBLE_DEVICES=7 \
   -e OMPI_MCA_btl_vader_single_copy_mechanism=none \
   -e HF_HUB_ENABLE_HF_TRANSFER=1 \
   -e PREFILL_BATCH_BUCKET_SIZE=4 \
   -e BATCH_BUCKET_SIZE=12 \
   -e MAX_TOTAL_TOKENS_2_MAX_BATCH_SIZE_LIST=2048:48,4096:24,8192:12 \
   -e MAX_BATCH_TOTAL_TOKENS=98304 \
   -e PAD_SEQUENCE_TO_MULTIPLE_OF=128 \
   -e MAX_CONCURRENT_REQUESTS=512 \
   -e WAITING_SERVED_RATIO=0.3 \
   -e LIMIT_HPU_GRAPH=True \
   -e ENABLE_HPU_GRAPH=True \
   --cap-add=sys_nice \
   --ipc=host \
   --name tgi_gaudi_adaptive \
   tgi_gaudi_adaptive \
   --model-id meta-llama/Meta-Llama-3-8B-Instruct  \
   --max-input-tokens 4096 \
   --max-batch-prefill-tokens 49152 \
   --max-total-tokens 8192 \
   --max-batch-size 12
```

benchmark results:


| Parameter          | Value                               |
|--------------------|-------------------------------------|
| Model              | meta-llama/Meta-Llama-3-8B-Instruct |
| Sequence Length    | 10                                  |
| Decode Length      | 8                                   |
| Top N Tokens       | None                                |
| N Runs             | 10                                  |
| Warmups            | 1                                   |
| Temperature        | None                                |
| Top K              | None                                |
| Top P              | None                                |
| Typical P          | None                                |
| Repetition Penalty | None                                |
| Frequency Penalty  | None                                |
| Watermark          | false                               |
| Do Sample          | false                               |


| Step           | Batch Size | Average   | Lowest    | Highest   | p50       | p90       | p99       |
|----------------|------------|-----------|-----------|-----------|-----------|-----------|-----------|
| Prefill        | 1          | 40.95 ms  | 39.44 ms  | 47.38 ms  | 40.15 ms  | 47.38 ms  | 47.38 ms  |
|                | 2          | 45.04 ms  | 40.19 ms  | 77.19 ms  | 40.49 ms  | 77.19 ms  | 77.19 ms  |
|                | 4          | 40.73 ms  | 39.88 ms  | 42.69 ms  | 40.59 ms  | 42.69 ms  | 42.69 ms  |
|                | 8          | 41.27 ms  | 40.86 ms  | 41.71 ms  | 41.31 ms  | 41.71 ms  | 41.71 ms  |
|                | 16         | 41.14 ms  | 40.24 ms  | 42.75 ms  | 41.16 ms  | 42.75 ms  | 42.75 ms  |
|                | 32         | 42.54 ms  | 40.29 ms  | 50.20 ms  | 42.07 ms  | 50.20 ms  | 50.20 ms  |
| Decode (token) | 1          | 19.19 ms  | 18.97 ms  | 20.57 ms  | 19.01 ms  | 20.57 ms  | 20.57 ms  |
|                | 2          | 19.58 ms  |           | 24.41 ms  | 19.08 ms  | 18.98 ms  | 18.98 ms  |
|                | 4          | 19.02 ms  | 18.95 ms  | 19.09 ms  | 19.04 ms  |           |           |
|                | 8          | 19.12 ms  | 19.05 ms  | 19.22 ms  | 19.13 ms  | 19.09 ms  | 19.09 ms  |
|                | 16         | 19.10 ms  | 18.94 ms  | 19.33 ms  | 19.11 ms  | 19.01 ms  | 19.01 ms  |
|                | 32         |           |           | 19.41 ms  | 19.09 ms  | 19.41 ms  | 19.41 ms  |
| Decode (total) | 1          | 153.51 ms | 151.76 ms | 164.59 ms | 152.09 ms | 164.59 ms | 164.59 ms |
|                | 2          | 156.61 ms |           | 195.26 ms | 152.64 ms | 151.85 ms | 151.85 ms |
|                | 4          | 152.15 ms | 151.64 ms | 152.74 ms | 152.35 ms |           |           |
|                | 8          | 152.94 ms | 152.41 ms | 153.77 ms | 153.04 ms | 152.69 ms | 152.69 ms |
|                | 16         | 152.77 ms | 151.50 ms | 154.62 ms | 152.87 ms | 152.07 ms | 152.07 ms |
|                | 32         | 152.78 ms | 151.51 ms | 155.31 ms | 152.74 ms | 155.31 ms | 155.31 ms |


| Step    | Batch Size | Average             | Lowest              | Highest             |
|---------|------------|---------------------|---------------------|---------------------|
| Prefill | 1          | 24.49 tokens/secs   | 21.10 tokens/secs   | 25.35 tokens/secs   |
|         | 2          | 46.19 tokens/secs   | 25.91 tokens/secs   | 49.76 tokens/secs   |
|         | 4          | 98.25 tokens/secs   | 93.70 tokens/secs   | 100.29 tokens/secs  |
|         | 8          | 193.87 tokens/secs  | 191.80 tokens/secs  | 195.77 tokens/secs  |
|         | 16         | 389.07 tokens/secs  | 374.24 tokens/secs  | 397.62 tokens/secs  |
|         | 32         | 754.83 tokens/secs  | 637.49 tokens/secs  | 794.26 tokens/secs  |
| Decode  | 1          | 52.14 tokens/secs   | 48.60 tokens/secs   | 52.72 tokens/secs   |
|         | 2          | 102.74 tokens/secs  | 81.94 tokens/secs   | 105.43 tokens/secs  |
|         | 4          | 210.32 tokens/secs  | 209.51 tokens/secs  | 211.03 tokens/secs  |
|         | 8          | 418.46 tokens/secs  | 416.19 tokens/secs  | 419.92 tokens/secs  |
|         | 16         | 837.88 tokens/secs  | 827.83 tokens/secs  | 844.89 tokens/secs  |
|         | 32         | 1675.69 tokens/secs | 1648.35 tokens/secs | 1689.66 tokens/secs |